Workaround for the issue in tabTitleFromTemplateId() method from Backend.Template js class.
For templates that are deep into the templates tree (admin>customize>edit templates),
the dhtmlXTreeObject.parentObject.id returned an array, while a string was expected.
As a result tabTitleFromTemplateId would die and not allow to edit the template.
dhtmlxtree is some proprietary library, so I won't try to fix it.