Fixing currency conversion in CustomerOrder::changeCurrency()
Added a line to reset the cached $toArrayData after each OrderedItem update, because otherwise OrderedItem::toArray() returns the price in the old currency.
This resulted in incorrect prices in the shopping cart when alternating between currencies. It showed the old currency price as discounted, and with the wrong currency sign.
A test was added to show the problem.