Update listPrices for products with variations.
Up till now if a user had a discount, the products with variations showed realPrice and listPrice, as the listPrice was shown with strikethrough formatting.
The problem was that when changing the selected variation, the realPrice was changed, but the listPrice was not. This update fixes this problem,
and now both prices are changed accordingly.