Fixing list prices on product details page in backend.
Even though only one currency was defined in a product, if the product had list price defined, all other currencies also calculated the list price, but not the base price.
As a result only the list price for a currency was filled, and when the form was submitted (presumably editing something else), the base price was saved as 0.
The base price is what the customer pays when adds the product in his shopping cart!

This fix makes sure that the list price is not filled if a currency was not defined for a product. This way Livecart automatically calculates the prices in the other currencies, based on the currency exchange rate.