Added mail queues for multi-threaded email sending.
Currently supports only MySQL and Semaphore queue, but can be extended to support all kind of queues.
You just need to add WhateverQueue.php file that extends QueueInterface.php inside model/queue folder.
All emails will be added to a Queue, and a cronjob(or many cronjobs) will get the queued emails,
and send them with the configured email transport system.

This results in faster newsletter sending, and the clients are not waiting on checkout/complete for
the server to connect to smtp server and send the emails.

Add as many threads as you need to your crontab like this:
* * * * * /usr/bin/php /public_html/storage/processQueues.php > /dev/null 2>&1
The more threads you have, the faster the emails will be sent, and the higher the
CPU load will be. So please use responsibly!